refactor(web): 폼 스키마 schema.ts로 분리
